+++ src/meltano/cli/invoke.py
@@ -4,6 +4,7 @@ import logging
 from . import cli
 from .params import project
 
+from meltano.core.plugin import PluginType
 from meltano.core.plugin_invoker import invoker_factory
 from meltano.core.config_service import ConfigService
 from meltano.core.tracking import GoogleAnalyticsTracker
@@ -11,16 +12,23 @@ from meltano.core.db import project_engine
 
 
 @cli.command(context_settings=dict(ignore_unknown_options=True))
+@click.option(
+    "--plugin-type", type=click.Choice(PluginType.cli_arguments()), default=None
+)
 @click.argument("plugin_name")
 @click.argument("plugin_args", nargs=-1, type=click.UNPROCESSED)
 @project(migrate=True)
-def invoke(project, plugin_name, plugin_args):
+def invoke(project, plugin_type, plugin_name, plugin_args):
+    plugin_type = PluginType.from_cli_argument(plugin_type) if plugin_type else None
+
     _, Session = project_engine(project)
 
     try:
         session = Session()
         config_service = ConfigService(project)
-        plugin = config_service.find_plugin(plugin_name)
+        plugin = config_service.find_plugin(
+            plugin_name, plugin_type=plugin_type, invokable=True
+        )
         service = invoker_factory(project, plugin, prepare_with_session=session)
         handle = service.invoke(*plugin_args)